Split our store into 2: one for each view
